move Ribbon component source out of demo folder
