+++ src/components/strip.js
